2월 1일 - 허민성
1. 보안
2. 카카오, 네이버, 구글 소셜 로그인
3. 검색 -동의어/ 한영검색

+++ src/main/java/com/project/springboot/dto/ArtistDTO.java
@@ -9,6 +9,6 @@ public class ArtistDTO {
     private String a_image;
     private String a_genres;
     private String a_debut_date;
-    private int a_followers;
+    private Long a_followers;
     private String a_spotify_id;
 }
